Rename workflow activity context and add an interface
