+++ examples/src/main/java/io/dapr/examples/workflows/continueasnew/CleanUpActivity.java
@@ -14,7 +14,7 @@ limitations under the License.
 package io.dapr.examples.workflows.continueasnew;
 
 import io.dapr.workflows.WorkflowActivity;
-import io.dapr.workflows.runtime.WorkflowActivityContext;
+import io.dapr.workflows.WorkflowActivityContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
