+++ sdk-tests/src/test/java/io/dapr/it/testcontainers/FirstActivity.java
@@ -14,7 +14,7 @@ limitations under the License.
 package io.dapr.it.testcontainers;
 
 import io.dapr.workflows.WorkflowActivity;
-import io.dapr.workflows.runtime.WorkflowActivityContext;
+import io.dapr.workflows.WorkflowActivityContext;
 
 public class FirstActivity implements WorkflowActivity {
