+++ sdk-tests/src/test/java/io/dapr/it/testcontainers/SecondActivity.java
@@ -14,7 +14,7 @@ limitations under the License.
 package io.dapr.it.testcontainers;
 
 import io.dapr.workflows.WorkflowActivity;
-import io.dapr.workflows.runtime.WorkflowActivityContext;
+import io.dapr.workflows.WorkflowActivityContext;
 
 public class SecondActivity implements WorkflowActivity {
