+++ sdk-workflows/src/main/java/io/dapr/workflows/WorkflowActivity.java
@@ -13,8 +13,6 @@ limitations under the License.
 
 package io.dapr.workflows;
 
-import io.dapr.workflows.runtime.WorkflowActivityContext;
-
 /**
  * Common interface for task activity implementations.
  *
