+++ sdk-workflows/src/main/java/io/dapr/workflows/WorkflowActivityContext.java
@@ -1,0 +1,7 @@
+package io.dapr.workflows;
+
+public interface WorkflowActivityContext {
+  String getName();
+
+  <T> T getInput(Class<T> targetType);
+}
