+++ sdk-workflows/src/test/java/io/dapr/workflows/runtime/WorkflowActivityWrapperTest.java
@@ -2,6 +2,7 @@ package io.dapr.workflows.runtime;
 
 import com.microsoft.durabletask.TaskActivityContext;
 import io.dapr.workflows.WorkflowActivity;
+import io.dapr.workflows.WorkflowActivityContext;
 import org.junit.Assert;
 import org.junit.Test;
